Guard against null values
Mapping the default configuration introduces `null` values that will
throw a console error if they get to the Input. To avoid the issue, we
use default values in `TextField` and `NumberField` that will avoid the
error while not affecting the desired effect of those `null`s: sending
the field back to the server even if it wasn't set.